refactor: Add transcription-filter-properties.cpp for managing filter properties
