+++ data/locale/en-US.ini
@@ -53,7 +53,7 @@ whisper_translate="Translate to English (Whisper)"
 buffer_size_msec="Buffer size (ms)"
 overlap_size_msec="Overlap size (ms)"
 suppress_sentences="Suppress sentences (each line)"
-translate_output="Translation output"
+translate_output="Output Destination"
 dtw_token_timestamps="DTW token timestamps"
 buffered_output="Buffered output (Experimental)"
 translate_model="Model"
